+++ inst/tinytest/test_ctrdata_other_functions.R
@@ -13,6 +13,20 @@ statusvalues <- list(
   "Lastvalue"   = c("34"))
 
 
+#### environment ####
+
+if (.Platform$OS.type != "windows") {
+  expect_error(
+    installCygwinWindowsDoInstall(),
+    "only for MS Windows")
+}
+if (.Platform$OS.type == "windows") {
+  expect_message(
+    installCygwinWindowsDoInstall(),
+  "cygwin is already installed")
+}
+
+
 #### dfMergeTwoVariablesRelevel ####
 
 # test
@@ -50,22 +64,43 @@ expect_error(
     colnames = 1:3),
   "Please provide exactly two column names.")
 
+# test
+expect_warning(
+  dfMergeTwoVariablesRelevel(
+    df = df,
+    varnames = c("var1", "var2")),
+  "Parameter varnames is deprecated, use colnames instead")
 
-#### ctrOpenSearchPagesInBrowser ####
-if (!at_home()) exit_file("Reason: not at_home")
-if (Sys.getenv("ON_APPVEYOR") != "") exit_file("Reason: on Appveyor")
-if (!check_internet()) exit_file("Reason: no internet connectivity")
+# test
+expect_error(
+  dfMergeTwoVariablesRelevel(
+    df = df,
+    colnames = c("var1", "var2"),
+    levelslist = 1:2),
+  "Need list for parameter 'levelslist'")
+
+
+#### ctrGetQueryUrlFromBrowser ####
 
 expect_equal(
   suppressWarnings(ctrGetQueryUrlFromBrowser(
-    "something_insensible")),
+    "ThisDoesNotExist")),
   NULL)
+
+# ctrGetQueryUrlFromBrowser(url = "type=Intr&age=0&intr=Drug&phase=0&phase=1&strd_e=12%2F31%2F2010")
+# ctrGetQueryUrlFromBrowser(url = "type=Intr&age=0&intr=Drug&phase=0&phase=1&strd_e=12%2F31%2F2010", "CTGOV")
+# ctrGetQueryUrlFromBrowser(url = "https://clinicaltrials.gov/ct2/results?type=Intr&cond=cancer&age=0")
+# ctrGetQueryUrlFromBrowser(url = "https://clinicaltrials.gov/ct2/results?type=Intr&cond=cancer&age=0", "CTGOV")
+# ctrGetQueryUrlFromBrowser(url = "https://clinicaltrials.gov/ct2/results?type=Intr&cond=cancer&age=0", "EUCTR")
+# ctrGetQueryUrlFromBrowser(url = "")
+# ctrGetQueryUrlFromBrowser(url = NA)
+# ctrGetQueryUrlFromBrowser(url = list())
 
 q <- "https://clinicaltrials.gov/ct2/results?type=Intr&cond=cancer&age=0"
 
 tmp_test <- suppressMessages(
   ctrGetQueryUrlFromBrowser(
-    content = q))
+    url = q))
 
 # test
 expect_true("data.frame" %in% class(tmp_test))
@@ -73,8 +108,25 @@ expect_true("data.frame" %in% class(tmp_test))
 # test
 expect_warning(
   ctrGetQueryUrlFromBrowser(
-    content = "ThisDoesNotExist"),
+    url = "ThisDoesNotExist"),
   "no clinical trial register search URL found")
+
+# test if query= is added
+expect_equal(
+  suppressMessages(
+    ctrGetQueryUrlFromBrowser(
+      url = "query=cancer&status=completed",
+      register = "EUCTR")),
+  suppressMessages(
+    ctrGetQueryUrlFromBrowser(
+      url = "cancer&status=completed",
+      register = "EUCTR"))
+)
+
+#### ctrOpenSearchPagesInBrowser ####
+
+if (!at_home()) exit_file("Reason: not at_home")
+if (!check_internet()) exit_file("Reason: no internet connectivity")
 
 # test
 expect_message(
@@ -91,7 +143,7 @@ q <- paste0("https://www.clinicaltrialsregister.eu/ctr-search/",
 
 tmp_test <- suppressMessages(
   ctrGetQueryUrlFromBrowser(
-    content = q))
+    url = q))
 
 # test
 expect_true("data.frame" %in% class(tmp_test))